:recyle: migrate test case to CustomViewTests
